requirements.txt
# ============================================================================
# Web Scraping & Data Collection
# ============================================================================
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
wikipedia==1.4.0
arxiv==2.1.0
lxml==4.9.3
feedparser==6.0.10

# ============================================================================
# Text Processing & NLP
# ============================================================================
nltk==3.8.1
pandas==2.1.1
numpy==1.24.3
spacy==3.7.2
textblob==0.17.1
language-tool-python==2.8

# ============================================================================
# LLM & RAG Framework
# ============================================================================
openai==1.10.0
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
pinecone-client==3.0.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# ============================================================================
# Embeddings & ML Models
# ============================================================================
transformers==4.35.2
torch==2.0.1
torchvision==0.15.2
scikit-learn==1.3.2
scipy==1.11.4

# ============================================================================
# Vector Database & Search
# ============================================================================
elasticsearch==8.11.0
redis==5.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# ============================================================================
# Web UI & Visualization
# ============================================================================
streamlit==1.28.0
plotly==5.17.0
matplotlib==3.8.0
seaborn==0.13.0
altair==5.1.2

# ============================================================================
# API Development
# ============================================================================
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pydantic-settings==2.1.0

# ============================================================================
# Utilities & Helpers
# ============================================================================
python-dotenv==1.0.0
tqdm==4.66.1
pyyaml==6.0.1
toml==0.10.2
python-dateutil==2.8.2

# ============================================================================
# Data Processing & Validation
# ============================================================================
jsonschema==4.20.0
validators==0.22.0
python-multipart==0.0.6

# ============================================================================
# Caching & Performance
# ============================================================================
cachetools==5.3.2
diskcache==5.6.3

# ============================================================================
# Monitoring & Logging
# ============================================================================
sentry-sdk==1.38.0
python-json-logger==2.0.7
structlog==23.2.0

# ============================================================================
# Data Pipeline & Scheduling
# ============================================================================
APScheduler==3.10.4
celery==5.3.4
python-rq==1.15.1

# ============================================================================
# Testing
# ============================================================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
responses==0.14.0

# ============================================================================
# Code Quality & Formatting
# ============================================================================
black==23.12.0
flake8==6.1.0
isort==5.13.2
mypy==1.7.1

# ============================================================================
# Document Processing (Optional but recommended)
# ============================================================================
PyPDF2==3.0.1
python-pptx==0.6.21
python-docx==0.8.11

# ============================================================================
# Advanced NLP Features (Optional)
# ============================================================================
spacy-lookups-data==1.0.3
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl

# ============================================================================
# Deployment & DevOps (Optional)
# ============================================================================
docker==6.1.0
python-json-logger==2.0.7
gunicorn==21.2.0

# ============================================================================
# Data Versioning & Experiment Tracking (Optional)
# ============================================================================
mlflow==2.10.2
wandb==0.16.1

# ============================================================================
# Visualization & Analytics (Optional)
# ============================================================================
dash==2.14.1
bokeh==3.3.0
kaleido==0.2.1